+++ CHANGELOG.md
@@ -7,5 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
-- Glenda parser
+- Great Lakes Environmental Database(Glenda) parser
   [GLM-110](https://opensource.ncsa.illinois.edu/jira/browse/GLM-110)
+- Great Lakes Fish Monitoring and Surveillance Program Parser
+  [GLM-112](https://opensource.ncsa.illinois.edu/jira/browse/GLM-112)